[feature added] Implemented multi-carousel effect on testimonial cards
